backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1

# AI and ML
groq==0.13.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Environment and configuration
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.9.0.post0
aiofiles==24.1.0

# Logging
loguru==0.7.3

# CORS
fastapi-cors==0.0.6